+++ Source/Anvil/Views/ConfigurationUI/EnvironmentEditViewModel.cs
@@ -1,7 +1,10 @@
 ﻿using System;
-using System.Collections.ObjectModel;
+using System.Collections.Generic;
+using System.Linq;
 using System.Reactive.Linq;
+using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 using Anvil.Framework.ComponentModel;
 using Anvil.Framework.MVVM;
@@ -15,11 +18,29 @@ using ReactiveUI;
 
 namespace Anvil.Views.ConfigurationUI
 {
+    public sealed class EnvironmentVariableViewModel : ReactiveObject
+    {
+        private bool mIsSelected;
+
+        public EnvironmentVariableViewModel(EnvironmentVariable model)
+        {
+            Model = model;
+        }
+
+        public bool IsSelected
+        {
+            get { return mIsSelected; }
+            set { this.RaiseAndSetIfChanged(ref mIsSelected, value); }
+        }
+
+        public EnvironmentVariable Model { get; }
+    }
+
     public sealed class EnvironmentEditViewModel : DisposableViewModel
     {
         private readonly Func<EnvironmentVariable, Task> mAddEnvVarAsync;
         private readonly Func<EnvironmentVariable, Task> mDeleteEnvVarAsync;
-        private readonly ReadOnlyObservableCollection<EnvironmentVariable> mEnvironmentVariables;
+        private readonly ObservableCollectionExtended<EnvironmentVariableViewModel> mEnvironmentVariables;
         private EnvironmentVariable mAddNewVariable = new EnvironmentVariable();
 
         public EnvironmentEditViewModel(
@@ -29,18 +50,32 @@ namespace Anvil.Views.ConfigurationUI
         {
             mAddEnvVarAsync = addEnvVarAsync;
             mDeleteEnvVarAsync = deleteEnvVarAsync;
+            mEnvironmentVariables = new ObservableCollectionExtended<EnvironmentVariableViewModel>();
 
             environmentVariables
                 .ObserveOnDispatcher()
-                .Sort(SortExpressionComparer<EnvironmentVariable>.Ascending(x => x.Key).ThenByAscending(x => x.Value))
-                .Bind(out mEnvironmentVariables)
+                .Transform(x => new EnvironmentVariableViewModel(x))
+                .DisposeMany()
+                .Sort(SortExpressionComparer<EnvironmentVariableViewModel>.Ascending(x => x.Model.Key).ThenByAscending(x => x.Model.Value))
+                .Bind(mEnvironmentVariables)
                 .Subscribe().TrackWith(Disposables);
 
             AddCommand = ReactiveCommand.Create();
             AddCommand.Subscribe(_AddNew).TrackWith(Disposables);
 
-            DeleteCommand = ReactiveCommandEx.Create<EnvironmentVariable>();
+            DeleteCommand = ReactiveCommandEx.Create<EnvironmentVariableViewModel>();
             DeleteCommand.Subscribe(_Delete).TrackWith(Disposables);
+
+            var hasSelectedVariables = mEnvironmentVariables
+                .ToObservableChangeSet()
+                .AddKey(x => x.Model.Key)
+                .TrueForAny(x => x.WhenAnyValue(y => y.IsSelected), x => x);
+
+            CopySelectedCommand = ReactiveCommand.Create(hasSelectedVariables);
+            CopySelectedCommand.Subscribe(_CopySelectedToClipboard).TrackWith(Disposables);
+
+            PasteCommand = ReactiveCommand.Create();
+            PasteCommand.Subscribe(_Paste).TrackWith(Disposables);
         }
 
         public ReactiveCommand<object> AddCommand { get; }
@@ -51,9 +86,13 @@ namespace Anvil.Views.ConfigurationUI
             private set { this.RaiseAndSetIfChanged(ref mAddNewVariable, value); }
         }
 
-        public ReactiveCommand<EnvironmentVariable> DeleteCommand { get; }
+        public ReactiveCommand<object> CopySelectedCommand { get; }
 
-        public ReadOnlyObservableCollection<EnvironmentVariable> EnvironmentVariables => mEnvironmentVariables;
+        public ReactiveCommand<EnvironmentVariableViewModel> DeleteCommand { get; }
+
+        public IEnumerable<EnvironmentVariableViewModel> EnvironmentVariables => mEnvironmentVariables;
+
+        public ReactiveCommand<object> PasteCommand { get; }
 
         private async void _AddNew(object _)
         {
@@ -63,9 +102,59 @@ namespace Anvil.Views.ConfigurationUI
             await mAddEnvVarAsync(envVar);
         }
 
-        private async void _Delete(EnvironmentVariable envVar)
+        private void _CopySelectedToClipboard(object _)
         {
-            await mDeleteEnvVarAsync(envVar);
+            var selectedItems = EnvironmentVariables.Where(x => x.IsSelected).ToArray();
+
+            var text = new StringBuilder();
+            foreach(var item in selectedItems)
+            {
+                // being naive and assuming Key won't include tabs or newlines, they shouldn't be addable via the UI
+                text.AppendLine($"{item.Model.Key}\t{item.Model.Value}");
+            }
+
+            Clipboard.SetText(text.ToString());
+        }
+
+        private async void _Delete(EnvironmentVariableViewModel envVar)
+        {
+            await mDeleteEnvVarAsync(envVar.Model);
+        }
+
+        private async void _Paste(object _)
+        {
+            var text = Clipboard.GetText();
+            var lines = text.Split(new[] {'\r', '\n'}, StringSplitOptions.RemoveEmptyEntries);
+
+            foreach(var line in lines)
+            {
+                var tabIndex = line.IndexOf('\t');
+                string key, value;
+                if(tabIndex < 0)
+                {
+                    key = line;
+                    value = "";
+                }
+                else
+                {
+                    key = line.Substring(0, tabIndex);
+                    value = line.Substring(tabIndex + 1);
+                }
+
+                var existingVar = mEnvironmentVariables.FirstOrDefault(x => key.Equals(x.Model.Key, StringComparison.OrdinalIgnoreCase));
+                if(existingVar != null)
+                {
+                    existingVar.Model.Value = value;
+                }
+                else
+                {
+                    await mAddEnvVarAsync(new EnvironmentVariable
+                    {
+                        Key = key,
+                        Value = value
+                    });
+                }
+            }
         }
     }
 }